Save application info file path and reuse it on launch
If the application info is parsed from a file we can later use it to launch the
application with the webapp launcher rather then creating a tempory file for it. In case
where we don't have a application info file we still create a tempory file for it and pass
it to the launcher.

Signed-off-by: Simon Busch <morphis@gravedo.de>